requirements_modern.txt
# Modern PyTorch Versions (Requires Model Code Adaptation)
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.0
